pyproject.toml
[project]
name = "secure-llm"
version = "0.1.0"
description = "Trajectory-aware safety control system for Large Language Models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Secure LLM Team"}
]

dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Async
    "httpx>=0.26.0",
    "aioredis>=2.0.1",
    "asyncpg>=0.29.0",
    
    # ML/AI
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    
    # Vector store
    "pymilvus>=2.3.0",
    
    # Probabilistic
    "pyro-ppl>=1.8.6",
    
    # Monitoring
    "prometheus-client>=0.19.0",
    "structlog>=23.3.0",
    
    # Config
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.9",
    "black>=23.12.0",
]

vllm = [
    "vllm>=0.2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
